server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Aritro Saha <29025984+AritroSaha10@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-dotenv = "^1.0.0"
pymongo = {extras = ["srv"], version = "^4.5.0"}
google-cloud-storage = "^2.10.0"
pydub = "^0.25.1"
google-cloud-speech = "^2.21.0"
assemblyai = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = 'server.app:app'
